+++ src/lib.rs
@@ -87,6 +87,8 @@ let files: Vec::<PathBuf> = Dowser::default()
 ```
 */
 
+#![forbid(unsafe_code)]
+
 #![warn(clippy::filetype_is_file)]
 #![warn(clippy::integer_division)]
 #![warn(clippy::needless_borrow)]